move non-ASM package out of the ASM package
