+++ src/main/java/witchinggadgets/client/ClientProxy.java
@@ -14,7 +14,6 @@ import thaumcraft.client.fx.ParticleEngine;
 import thaumcraft.client.fx.particles.FXEssentiaTrail;
 import thaumcraft.client.fx.particles.FXWisp;
 import travellersgear.api.TravellersGearAPI;
-import witchinggadgets.asm.pouch.GuiPatchedFocusPouch;
 import witchinggadgets.client.fx.EntityFXSweat;
 import witchinggadgets.client.gui.GuiBag;
 import witchinggadgets.client.gui.GuiCloakBag;
@@ -60,6 +59,7 @@ import witchinggadgets.common.blocks.tiles.TileEntityTerraformFocus;
 import witchinggadgets.common.blocks.tiles.TileEntityTerraformer;
 import witchinggadgets.common.blocks.tiles.TileEntityWallMirror;
 import witchinggadgets.common.items.EntityItemReforming;
+import witchinggadgets.common.pouch.GuiPatchedFocusPouch;
 import witchinggadgets.common.util.WGKeyHandler;
 import baubles.api.BaublesApi;
 import cpw.mods.fml.client.FMLClientHandler;
