+++ src/main/java/witchinggadgets/common/CommonProxy.java
@@ -6,7 +6,6 @@ import net.minecraft.world.World;
 import net.minecraftforge.common.util.ForgeDirection;
 
 import travellersgear.api.TravellersGearAPI;
-import witchinggadgets.asm.pouch.ContainerPatchedFocusPouch;
 import witchinggadgets.common.blocks.tiles.TileEntityCuttingTable;
 import witchinggadgets.common.blocks.tiles.TileEntityLabelLibrary;
 import witchinggadgets.common.blocks.tiles.TileEntitySpinningWheel;
@@ -17,6 +16,7 @@ import witchinggadgets.common.gui.ContainerLabelLibrary;
 import witchinggadgets.common.gui.ContainerPrimordialGlove;
 import witchinggadgets.common.gui.ContainerSpinningWheel;
 import witchinggadgets.common.gui.ContainerVoidBag;
+import witchinggadgets.common.pouch.ContainerPatchedFocusPouch;
 import baubles.api.BaublesApi;
 import cpw.mods.fml.common.network.IGuiHandler;
