+++ src/main/java/witchinggadgets/common/pouch/GuiPatchedFocusPouch.java
@@ -1,4 +1,4 @@
-package witchinggadgets.asm.pouch;
+package witchinggadgets.common.pouch;
 
 import net.minecraft.client.gui.inventory.GuiContainer;
 import net.minecraft.entity.player.InventoryPlayer;
